test: refactor BasePage
