--- tests/src/test/java/selenium/BaseTest.java
@@ -1,43 +1,0 @@
-package selenium;
-
-import com.github.javafaker.Faker;
-import io.github.bonigarcia.wdm.WebDriverManager;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.edge.EdgeDriver;
-import org.openqa.selenium.edge.EdgeOptions;
-import org.openqa.selenium.support.ui.WebDriverWait;
-
-import java.time.Duration;
-
-public class BaseTest {
-
-    protected WebDriver driver;
-    protected WebDriverWait wait;
-    protected Faker faker;
-
-    @BeforeEach
-    public void setUp(){
-        WebDriverManager.edgedriver().setup();
-        EdgeOptions options = new EdgeOptions();
-        options.addArguments("user-agent=Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.3");
-        options.addArguments("--remote-allow-origins=*");
-        options.addArguments("--disable-web-security");
-        driver = new EdgeDriver(options);
-        driver.manage().window().maximize();
-        driver.get("https://site-teste-renatoxd152s-projects.vercel.app");
-
-        wait = new WebDriverWait(driver, Duration.ofSeconds(3));
-
-        faker = new Faker();
-    }
-
-    @AfterEach
-    public void tearDown(){
-        if (driver != null)
-            driver.quit();
-    }
-
-}
